requirements.txt
python-decouple
RestrictedPython
openai